requirements.txt
anthropic
python-dotenv
tavily-python
requests
beautifulsoup4
numpy
voyageai
langgraph